nextcloud hub10 update

+++ nextcloud/apps/groupfolders/lib/Controller/FolderController.php
@@ -8,14 +8,18 @@
 namespace OCA\GroupFolders\Controller;
 
 use OC\AppFramework\OCS\V1Response;
+use OCA\GroupFolders\Attribute\RequireGroupFolderAdmin;
 use OCA\GroupFolders\Folder\FolderManager;
 use OCA\GroupFolders\Mount\MountProvider;
 use OCA\GroupFolders\ResponseDefinitions;
 use OCA\GroupFolders\Service\DelegationService;
 use OCA\GroupFolders\Service\FoldersFilter;
 use OCP\AppFramework\Http;
+use OCP\AppFramework\Http\Attribute\FrontpageRoute;
+use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\PasswordConfirmationRequired;
 use OCP\AppFramework\Http\DataResponse;
+use OCP\AppFramework\OCS\OCSBadRequestException;
 use OCP\AppFramework\OCS\OCSForbiddenException;
 use OCP\AppFramework\OCS\OCSNotFoundException;
 use OCP\AppFramework\OCSController;
@@ -33,37 +37,30 @@ use OCP\IUserSession;
  * @psalm-import-type InternalFolderOut from FolderManager
  */
 class FolderController extends OCSController {
-	private FolderManager $manager;
-	private MountProvider $mountProvider;
-	private IRootFolder $rootFolder;
-	private ?IUser $user = null;
-	private FoldersFilter $foldersFilter;
-	private DelegationService $delegationService;
-	private IGroupManager $groupManager;
+	private ?IUser $user;
+
+	protected const ALLOWED_ORDER_BY = [
+		'mount_point',
+		'quota',
+		'groups',
+		'acl',
+	];
 
 	public function __construct(
 		string $AppName,
 		IRequest $request,
-		FolderManager $manager,
-		MountProvider $mountProvider,
-		IRootFolder $rootFolder,
+		private FolderManager $manager,
+		private MountProvider $mountProvider,
+		private IRootFolder $rootFolder,
 		IUserSession $userSession,
-		FoldersFilter $foldersFilter,
-		DelegationService $delegationService,
-		IGroupManager $groupManager,
+		private FoldersFilter $foldersFilter,
+		private DelegationService $delegationService,
+		private IGroupManager $groupManager,
 	) {
 		parent::__construct($AppName, $request);
-		$this->foldersFilter = $foldersFilter;
-		$this->manager = $manager;
-		$this->mountProvider = $mountProvider;
-		$this->rootFolder = $rootFolder;
 		$this->user = $userSession->getUser();
 
-		$this->registerResponder('xml', function ($data): V1Response {
-			return $this->buildOCSResponseXML('xml', $data);
-		});
-		$this->delegationService = $delegationService;
-		$this->groupManager = $groupManager;
+		$this->registerResponder('xml', fn (DataResponse $data): V1Response => $this->buildOCSResponseXML('xml', $data));
 	}
 
 	/**
@@ -73,6 +70,10 @@ class FolderController extends OCSController {
 	 * @return null|GroupFoldersFolder
 	 */
 	private function filterNonAdminFolder(array $folder): ?array {
+		if ($this->user === null) {
+			return null;
+		}
+
 		$userGroups = $this->groupManager->getUserGroupIds($this->user);
 		$folder['groups'] = array_filter($folder['groups'], static fn (string $group): bool => in_array($group, $userGroups, true), ARRAY_FILTER_USE_KEY);
 		$folder['group_details'] = array_filter($folder['group_details'], static fn (string $group): bool => in_array($group, $userGroups, true), ARRAY_FILTER_USE_KEY);
@@ -90,22 +91,31 @@ class FolderController extends OCSController {
 	private function formatFolder(array $folder): array {
 		// keep compatibility with the old 'groups' field
 		$folder['group_details'] = $folder['groups'];
-		$folder['groups'] = array_map(function (array $group) {
-			return $group['permissions'];
-		}, $folder['groups']);
+		$folder['groups'] = array_map(fn (array $group): int => $group['permissions'], $folder['groups']);
+
 		return $folder;
 	}
 
 	/**
 	 * Gets all Groupfolders
-	 * @NoAdminRequired
+	 *
 	 * @param bool $applicable Filter by applicable groups
+	 * @param non-negative-int $offset Number of items to skip.
+	 * @param ?positive-int $limit Number of items to return.
+	 * @param null|'mount_point'|'quota'|'groups'|'acl' $orderBy The key to order by
 	 * @return DataResponse<Http::STATUS_OK, array<string, GroupFoldersFolder>, array{}>
 	 * @throws OCSNotFoundException Storage not found
+	 * @throws OCSBadRequestException Wrong limit used
 	 *
 	 * 200: Groupfolders returned
 	 */
-	public function getFolders(bool $applicable = false): DataResponse {
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'GET', url: '/folders')]
+	public function getFolders(bool $applicable = false, int $offset = 0, ?int $limit = null, ?string $orderBy = 'mount_point'): DataResponse {
+		if ($limit !== null && $limit <= 0) {
+			throw new OCSBadRequestException('The limit must be greater than 0.');
+		}
+
 		$storageId = $this->getRootFolderStorageId();
 		if ($storageId === null) {
 			throw new OCSNotFoundException();
@@ -116,30 +126,79 @@ class FolderController extends OCSController {
 			// Make them string-indexed for OpenAPI JSON output
 			$folders[(string)$id] = $this->formatFolder($folder);
 		}
+
+		$orderBy = in_array($orderBy, self::ALLOWED_ORDER_BY, true)
+			? $orderBy
+			: 'mount_point';
+
+		// in case of equal orderBy value always fall back to the mount_point - same as on the frontend
+		/**
+		 * @var GroupFoldersFolder $a
+		 * @var GroupFoldersFolder $b
+		 */
+		uasort($folders, function (array $a, array $b) use ($orderBy) {
+			if ($orderBy === 'groups') {
+				if (($value = count($a['groups']) - count($b['groups'])) !== 0) {
+					return $value;
+				}
+			} else {
+				if (($value = strcmp((string)($a[$orderBy] ?? ''), (string)($b[$orderBy] ?? ''))) !== 0) {
+					return $value;
+				}
+			}
+
+			// fallback to mount_point
+			if (($value = strcmp($a['mount_point'] ?? '', $b['mount_point'])) !== 0) {
+				return $value;
+			}
+
+			// fallback to id
+			return $a['id'] - $b['id'];
+		});
+
 		$isAdmin = $this->delegationService->isAdminNextcloud() || $this->delegationService->isDelegatedAdmin();
 		if ($isAdmin && !$applicable) {
+			// If only the default values are provided the pagination can be skipped.
+			if ($offset !== 0 || $limit !== null) {
+				$folders = array_slice($folders, $offset, $limit, true);
+			}
+
 			return new DataResponse($folders);
 		}
+
 		if ($this->delegationService->hasOnlyApiAccess()) {
 			$folders = $this->foldersFilter->getForApiUser($folders);
 		}
+
 		if ($applicable || !$this->delegationService->hasApiAccess()) {
 			$folders = array_filter(array_map($this->filterNonAdminFolder(...), $folders));
 		}
+
+		// If only the default values are provided the pagination can be skipped.
+		if ($offset !== 0 || $limit !== null) {
+			$folders = array_slice($folders, $offset, $limit, true);
+		}
+
 		return new DataResponse($folders);
 	}
 
 	/**
 	 * Gets a Groupfolder by ID
-	 * @NoAdminRequired
+	 *
 	 * @param int $id ID of the Groupfolder
-	 * @return DataResponse<Http::STATUS_OK, GroupFoldersFolder, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
+	 * @return DataResponse<Http::STATUS_OK, GroupFoldersFolder, array{}>
 	 * @throws OCSNotFoundException Groupfolder not found
 	 *
 	 * 200: Groupfolder returned
 	 */
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'GET', url: '/folders/{id}')]
 	public function getFolder(int $id): DataResponse {
 		$storageId = $this->getRootFolderStorageId();
+		if ($storageId === null) {
+			throw new OCSNotFoundException();
+		}
+
 		$folder = $this->manager->getFolder($id, $storageId);
 		if ($folder === null) {
 			throw new OCSNotFoundException();
@@ -148,10 +207,11 @@ class FolderController extends OCSController {
 
 		if (!$this->delegationService->hasApiAccess()) {
 			$folder = $this->filterNonAdminFolder($folder);
-			if (!$folder) {
-				return new DataResponse([], Http::STATUS_NOT_FOUND);
+			if ($folder === null) {
+				throw new OCSNotFoundException();
 			}
 		}
+
 		return new DataResponse($folder);
 	}
 
@@ -163,10 +223,12 @@ class FolderController extends OCSController {
 		if ($storageId === null) {
 			return new DataResponse([], Http::STATUS_NOT_FOUND);
 		}
+
 		$folder = $this->manager->getFolder($id, $storageId);
-		if ($folder === false) {
+		if ($folder === null) {
 			return new DataResponse([], Http::STATUS_NOT_FOUND);
 		}
+
 		return null;
 	}
 
@@ -176,8 +238,7 @@ class FolderController extends OCSController {
 
 	/**
 	 * Add a new Groupfolder
-	 * @RequireGroupFolderAdmin
-	 * @NoAdminRequired
+	 *
 	 * @param string $mountpoint Mountpoint of the new Groupfolder
 	 * @return DataResponse<Http::STATUS_OK, GroupFoldersFolder, array{}>
 	 * @throws OCSNotFoundException Groupfolder not found
@@ -185,19 +246,28 @@ class FolderController extends OCSController {
 	 * 200: Groupfolder added successfully
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders')]
 	public function addFolder(string $mountpoint): DataResponse {
-		$id = $this->manager->createFolder(trim($mountpoint));
-		$folder = $this->manager->getFolder($id, $this->rootFolder->getMountPoint()->getNumericStorageId());
-		if ($folder === false) {
+
+		$storageId = $this->rootFolder->getMountPoint()->getNumericStorageId();
+		if ($storageId === null) {
 			throw new OCSNotFoundException();
 		}
+
+		$id = $this->manager->createFolder(trim($mountpoint));
+		$folder = $this->manager->getFolder($id, $storageId);
+		if ($folder === null) {
+			throw new OCSNotFoundException();
+		}
+
 		return new DataResponse($this->formatFolder($folder));
 	}
 
 	/**
 	 * Remove a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
 	 * @throws OCSNotFoundException Groupfolder not found
@@ -206,21 +276,29 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'DELETE', url: '/folders/{id}')]
 	public function removeFolder(int $id): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$folder = $this->mountProvider->getFolder($id);
+		if ($folder === null) {
+			throw new OCSNotFoundException();
+		}
+
 		$folder->delete();
 		$this->manager->removeFolder($id);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Set the mount point of a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param string $mountPoint New mount point path
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>
@@ -228,6 +306,9 @@ class FolderController extends OCSController {
 	 * 200: Mount point changed successfully
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'PUT', url: '/folders/{id}')]
 	public function setMountPoint(int $id, string $mountPoint): DataResponse {
 		$this->manager->renameFolder($id, trim($mountPoint));
 		return new DataResponse(['success' => true]);
@@ -235,8 +316,7 @@ class FolderController extends OCSController {
 
 	/**
 	 * Add access of a group for a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param string $group Group to add access for
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
@@ -245,19 +325,23 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders/{id}/groups')]
 	public function addGroup(int $id, string $group): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->addApplicableGroup($id, $group);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Remove access of a group from a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param string $group Group to remove access from
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
@@ -266,19 +350,23 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'DELETE', url: '/folders/{id}/groups/{group}', requirements: ['group' => '.+'])]
 	public function removeGroup(int $id, string $group): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->removeApplicableGroup($id, $group);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Set the permissions of a group for a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param string $group Group for which the permissions will be set
 	 * @param int $permissions New permissions
@@ -288,19 +376,23 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders/{id}/groups/{group}', requirements: ['group' => '.+'])]
 	public function setPermissions(int $id, string $group, int $permissions): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->setGroupPermissions($id, $group, $permissions);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Updates an ACL mapping
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param string $mappingType Type of the ACL mapping
 	 * @param string $mappingId ID of the ACL mapping
@@ -311,19 +403,23 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders/{id}/manageACL')]
 	public function setManageACL(int $id, string $mappingType, string $mappingId, bool $manageAcl): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->setManageACL($id, $mappingType, $mappingId, $manageAcl);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Set a new quota for a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param int $quota New quota in bytes
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
@@ -332,19 +428,23 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders/{id}/quota')]
 	public function setQuota(int $id, int $quota): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->setFolderQuota($id, $quota);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Toggle the ACL for a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param bool $acl Whether ACL should be enabled or not
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
@@ -353,19 +453,23 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders/{id}/acl')]
 	public function setACL(int $id, bool $acl): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->setFolderACL($id, $acl);
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Rename a Groupfolder
-	 * @NoAdminRequired
-	 * @RequireGroupFolderAdmin
+	 *
 	 * @param int $id ID of the Groupfolder
 	 * @param string $mountpoint New Mountpoint of the Groupfolder
 	 * @return DataResponse<Http::STATUS_OK, array{success: true}, array{}>|DataResponse<Http::STATUS_NOT_FOUND, list<empty>, array{}>
@@ -374,22 +478,22 @@ class FolderController extends OCSController {
 	 * 404: Groupfolder not found
 	 */
 	#[PasswordConfirmationRequired]
+	#[RequireGroupFolderAdmin]
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'POST', url: '/folders/{id}/mountpoint')]
 	public function renameFolder(int $id, string $mountpoint): DataResponse {
 		$response = $this->checkFolderExists($id);
 		if ($response) {
 			return $response;
 		}
+
 		$this->manager->renameFolder($id, trim($mountpoint));
+
 		return new DataResponse(['success' => true]);
 	}
 
 	/**
 	 * Overwrite response builder to customize xml handling to deal with spaces in folder names
-	 *
-	 * @param string $format json or xml
-	 * @param DataResponse $data the data which should be transformed
-	 * @return \OC\AppFramework\OCS\V1Response
-	 * @since 8.1.0
 	 */
 	private function buildOCSResponseXML(string $format, DataResponse $data): V1Response {
 		/** @var array $folderData */
@@ -399,9 +503,11 @@ class FolderController extends OCSController {
 			$folderData = $this->folderDataForXML($folderData);
 		} elseif (is_array($folderData) && count($folderData) && isset(current($folderData)['id'])) {
 			// folder list
-			$folderData = array_map([$this, 'folderDataForXML'], $folderData);
+			$folderData = array_map($this->folderDataForXML(...), $folderData);
 		}
+
 		$data->setData($folderData);
+
 		return new V1Response($data, $format);
 	}
 
@@ -417,12 +523,13 @@ class FolderController extends OCSController {
 				'@type' => $group['type'],
 			];
 		}
+
 		return $data;
 	}
 
 	/**
 	 * Searches for matching ACL mappings
-	 * @NoAdminRequired
+	 *
 	 * @param int $id The ID of the Groupfolder
 	 * @param string $search String to search by
 	 * @return DataResponse<Http::STATUS_OK, array{users: list<GroupFoldersUser>, groups: list<GroupFoldersGroup>, circles: list<GroupFoldersCircle>}, array{}>
@@ -430,14 +537,21 @@ class FolderController extends OCSController {
 	 *
 	 * 200: ACL Mappings returned
 	 */
+	#[NoAdminRequired]
+	#[FrontpageRoute(verb: 'GET', url: '/folders/{id}/search')]
 	public function aclMappingSearch(int $id, string $search = ''): DataResponse {
 		$users = $groups = $circles = [];
+
+		if ($this->user === null) {
+			throw new OCSForbiddenException();
+		}
 
 		if ($this->manager->canManageACL($id, $this->user) === true) {
 			$groups = $this->manager->searchGroups($id, $search);
 			$users = $this->manager->searchUsers($id, $search);
 			$circles = $this->manager->searchCircles($id, $search);
 		}
+
 		return new DataResponse([
 			'users' => $users,
 			'groups' => $groups,
